remove ipython notebook template
the new direction for the prework is to have it all
submitted through the github interface, and it is
probably better to have just text rather than the ipynb
format, which would have to be copied and pasted
wholesale, most likely - and could be confusing.

also, we're thinking of changing the required exercises,
so the current selection identified in the template may
no longer be applicable.

ping @reshama @lauramontemayor